requirements.txt
pytest-cov == 2.7.1